pyproject.toml
[project]
name = "api-metatrader5"
version = "0.1.0"
description = "Private FastAPI gateway for MetaTrader 5 quotes and future order routing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
    "pydantic-settings>=2.8,<3.0",
    "MetaTrader5>=5.0.5388",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28,<1.0",
    "pytest>=8.3,<9.0",
    "ruff>=0.11,<1.0",
]

[project.scripts]
api-metatrader5 = "api_metatrader5.main:main"

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api_metatrader5"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
